crawler/requirements.txt
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
aiosqlite==0.22.1
alphashape==1.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
beautifulsoup4==4.14.3
brotli==1.2.0
certifi==2026.2.25
cffi==2.0.0
chardet==7.4.3
charset-normalizer==3.4.7
click==8.3.2
click-log==0.4.0
colorama==0.4.6
Crawl4AI==0.8.6
cryptography==46.0.7
cssselect==1.4.0
distro==1.9.0
fake-useragent==2.2.0
fastuuid==0.14.0
filelock==3.29.0
frozenlist==1.8.0
fsspec==2026.3.0
greenlet==3.4.0
h11==0.16.0
h2==4.3.0
hf-xet==1.4.3
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.11.0
humanize==4.15.0
hyperframe==6.1.0
idna==3.11
importlib_metadata==9.0.0
Jinja2==3.1.6
jiter==0.14.0
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
lark==1.3.1
lxml==5.4.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
multidict==6.7.1
networkx==3.6.1
nltk==3.9.4
numpy==2.4.4
openai==2.32.0
packaging==26.1
patchright==1.58.2
pillow==12.2.0
playwright==1.58.0
playwright-stealth==2.0.3
propcache==0.4.1
psutil==7.2.2
psycopg2-binary==2.9.12
pycparser==3.0
pydantic==2.13.3
pydantic_core==2.46.3
pyee==13.0.1
Pygments==2.20.0
pyOpenSSL==26.0.0
python-dotenv==1.2.2
PyYAML==6.0.3
rank-bm25==0.2.2
referencing==0.37.0
regex==2026.4.4
requests==2.33.1
rich==15.0.0
rpds-py==0.30.0
rtree==1.4.1
scipy==1.17.1
shapely==2.1.2
shellingham==1.5.4
sniffio==1.3.1
snowballstemmer==2.2.0
soupsieve==2.8.3
tiktoken==0.12.0
tokenizers==0.22.2
tqdm==4.67.3
trimesh==4.11.5
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
unclecode-litellm==1.81.13
urllib3==2.6.3
xxhash==3.6.0
yarl==1.23.0
zipp==3.23.1
